Rounding decimal values to two decimal places.

+++ ABMVantage.Data/Models/OccRevenueByProduct.cs
@@ -35,7 +35,7 @@
         public string Fiscal { get; set; }
         public decimal Occupancy { get; set; }
         public DateTime FirstDayOfMonth { get; internal set; }
-        public int OccupancyPercentage { get; internal set; }
+        public decimal OccupancyPercentage { get; internal set; }
     }
 
     public class YearlyOccupancyGroupedResult
